+++ RELEASES.md
@@ -1,5 +1,10 @@
 # Releases
 
+# 1.0.0 Promptzl Stable Version
+
+### Bugfixes
+ - Throw error if empty labels in verbalizer are passed.
+ - Add error if prompt is too long upon initialization
 
 ## 0.9.3